extraction of all ni modules done
